TASK-37339 Fix Cloud Drive button display in Attachments drawer Defining a module outside Vue and use it afterward in a Vue component leads to loosing context of Vue and thus instability in displaying the Cloud drive button. This PR will move the definition of Activity attachments extension to Vue context
