+++ apps/portlet-clouddrives/src/main/webapp/vue-app/cloudStorage/components/CloudStorage.vue
@@ -50,7 +50,7 @@
 </template>
 
 <script>
-import { isCloudDriveEnabled } from "../../cloudDriveService";
+import { isCloudDriveEnabled } from "../js/cloudDriveService";
 
 export default {
   data: function() {
