+++ apps/portlet-clouddrives/src/main/webapp/vue-app/cloudStorage/js/cloudDriveService.js
@@ -1,0 +1,11 @@
+export function isCloudDriveEnabled() {
+  return fetch(`${eXo.env.portal.context}/${eXo.env.portal.rest}/clouddrive/features/status/enabled`,{
+    credentials: "include",
+    method: "GET",
+  }).then((resp) => {
+    if (resp && resp.ok) {
+      return resp.json();
+    }
+  });
+}
+
