+++ apps/portlet-clouddrives/src/main/webapp/vue-app/connectCloudDrive/main.js
@@ -14,23 +14,14 @@ const lang = typeof eXo !== "undefined" ? eXo.env.portal.language : "en";
 // should expose the locale resources as REST API
 const url = `/portal/rest/i18n/bundle/locale.clouddrive.CloudDrive-${lang}.json`;
 
-let connectCloudApp;
-export function init() {
-  // getting locale resources
-  exoi18n.loadLanguageAsync(lang, url).then(i18n => {
-    // init Vue app when locale resources are ready
-    connectCloudApp = new Vue({
-      render: function(createElement) {
-        return createElement(ConnectCloudDrive, { props: { showCloudDrawer: true }});
-      },
-      i18n,
-      vuetify
-    }).$mount("#connectCloudApp");
-  });
-}
-
-export function destroy() {
-  if (connectCloudApp) {
-    connectCloudApp.$destroy();
-  }
-}
+// getting locale resources
+exoi18n.loadLanguageAsync(lang, url).then(i18n => {
+  // init Vue app when locale resources are ready
+  new Vue({
+    render: function(createElement) {
+      return createElement(ConnectCloudDrive, { props: { showCloudDrawer: true }});
+    },
+    i18n,
+    vuetify
+  }).$mount("#connectCloudApp");
+});
