+++ ecm-wcm-extension/src/main/webapp/attachments-selector/components/ExoFoldersFilesSelector.vue
@@ -362,6 +362,7 @@ export default {
       this.errorMessage= `${this.$t('attachments.fetchFoldersAndFiles.error')}`;
       this.showErrorMessage = true;
     });
+    document.addEventListener('extension-AttachmentsComposer-attachments-composer-action-updated', () => this.attachmentsComposerActions = getAttachmentsComposerExtensions());
     this.attachmentsComposerActions = getAttachmentsComposerExtensions();
   },
   methods: {
