+++ ecm-wcm-extension/src/main/webapp/attachments-selector/extension.js
@@ -17,7 +17,7 @@ export function getAttachmentsComposerExtensions() {
 
 export function executeExtensionAction(extension, component) {
   if(extension.hasOwnProperty('onExecute') && isFunction(extension.onExecute)) {
-    extension.onExecute(component); // it will execute code inside onExecute() defined in configuration file
+    extension.onExecute.apply(component, [component]); // it will execute code inside onExecute() defined in configuration file
   }
 }
